This PR adds performance optimizations for GPU accelerated token generation, mostly benefiting fast GPUs like the RTX 3090. Performance optimizations can be enabled via the options `LLAMA_CUDA_BY=2` and `LLAMA_CUDA_UNROLL=1` (make) or `LLAMA_CUDA_UNROLL=ON` (cmake) at compile time. **These options degrade performance on my GTX 1070.** Build instructions (Linux):

```
git clone https://github.com/JohannesGaessler/llama.cpp llama.cpp-johannesgaessler
cd llama.cpp-johannesgaessler                               
git fetch
git switch dfyz-xor-hack
make LLAMA_CUBLAS=1 LLAMA_CUDA_BY=2 LLAMA_CUDA_UNROLL=1
```

### Implementation details

* As suggested by @dfyz in a previous PR I have eliminated the shared memory from my CUDA kernel and am using black xor magic to sum up the partial sums at the ends. This was universally faster on my RTX 3090 and my GTX 1070. But because this breaks HIP compatibility for [this PR](https://github.com/ggerganov/llama.cpp/pull/1087) it is only being used if `GGML_USE_HIPBLAS` is not defined.
* Also suggested by @dfyz : larger blocks for the CUDA kernels by assigning two rows to each block. The option `LLAMA_CUDA_BY` sets the number of rows per block. On my RTX 3090 setting this option to 2 is faster but higher values have slightly worse performance. On my GTX 1070 a value of 2 or higher causes performance degradation.
* ~~Loop unrolling: The matrices used in llama.cpp always have the same size. So the loops used during inference can be unrolled if the compiler is told how large the matrices are. This is done via moving `ncols` from a regular argument to a template argument and adding a switch statement for the various matrix sizes (8 in total). On my RTX 3090 this is faster but on my GTX 1070 it's slower. **Enabling this option significantly increases compile time.**~~
* Larger blocks in x direction: The option `LLAMA_CUDA_BX` can be set to determine the block size in x direction. Default value is 32, 64 was faster on RTX 3090.

### Results

For the RTX 3090 I used `LLAMA_CUDA_BY=2 LLAMA_CUDA_UNROLL=1`, for the GTX 1070 I did **not** use these options.

| GPU      | Model    | ms/t master | ms/t PR | ms/t no unroll |
|----------|----------|-------------|---------|----|
| RTX 3090 | 7b q4_0  |       23.49 |   20.36 | 21.04 |
| RTX 3090 | 7b q8_0  |       24.54 |   22.05 | - |
| RTX 3090 | 13b q4_0 |       37.95 |   32.85 | 34.04 |
| RTX 3090 | 33b q4_0 |       83.15 |   69.76 | 73.46 |
| GTX 1070 | 7b q4_0  |       69.20 |   67.80 | - |
| GTX 1070 | 13b q4_0 |      141.30 |  139.39 | - |