@SlyEcho 2023-05-19
> But because this breaks HIP compatibility for https://github.com/ggerganov/llama.cpp/pull/1087 it is only being used if GGML_USE_HIPBLAS is not defined.

I would not worry about HIP right now, it is still in draft, and this is one of the reasons, it's not sure if it is fully compatible with CUDA (so far it has been).

I was able to solve the issue with this kind of change:
```diff
- __shfl_xor_sync(0xffffffff, tmp, mask, 32)
+ __shfl_xor(tmp, mask, 32)
```

Well actually, the 32 I changed to 64, etc, because of the different warp size. It would be nice if it were a define or something, actually.