@bluefireexplosion 2023-05-19
Excellent work! Interesting analytics:

- 7B shows a 15.3% speedup in ms/t
- 13B shows a 15.5% speedup in ms/t
- 33B shows a 19.1% speedup in ms/t

It would be interesting to see if the performance gain for 65B follows the trend of slight percentage increases as the parameter count increases.