@ggerganov 2023-05-20
The unroll indeed makes the compilation super long. Not sure we want to support it.
Couldn't you achieve the same thing via:

```c++
for (int i = 0; i < n; ++i) {
...
}
```

replace with:

```c++
assert(n % 32 == 0);
for (int io = 0; io < n; io += 32) {
    #pragma unroll
    for (int i = 0; i < 32; ++i) {
    ....
    }
}
```

Or something along these lines