@JohannesGaessler 2023-05-20
The unroll makes the compilation longer but honestly I don't care about 2 minutes longer compilation if it means I get a few % more performance.