@JohannesGaessler 2023-05-20
I've pushed an alternative version to [this branch](https://github.com/JohannesGaessler/llama.cpp/tree/dfyz-xor-hack-alt-unroll). The value for `istride` cannot be pushed very high; the value on the branch is the max value for 33b that is bug free. Notably unrolling the inner loop does nothing to improve performance. Unrolling both the inner and the outer loop gives you the same performance and compile time as this PR. I would argue that it's fundamentally impossible to get the performance uplift without the increase in compile time in the first place: the compiler spends that time reordering and optimizing the code which was previously impossible due to the loop.