@SlyEcho 2023-05-20
I am gonna try to bring #1087 into compatibility with this.

The shuffle function seems the biggest hurdle right now, but it doesn't [seem impossible](https://github.com/ROCm-Developer-Tools/HIP/issues/1491#issuecomment-1256171886).