@SlyEcho 2023-05-20
For AMD this is roughly 20% faster:
```cuda
#define GGML_CUDA_DMMV_BLOCK_X 64 // dmmv = dequantize_mul_mat_vec
// ...
for (int mask = GGML_CUDA_DMMV_BLOCK_X/2; mask > 0; mask >>= 1) {
    tmp += __shfl_xor_sync(0xffffffff, tmp, mask, GGML_CUDA_DMMV_BLOCK_X);
}
``` 

And `LLAMA_CUDA_BY=4 LLAMA_CUDA_UNROLL=ON` adds another couple percents. But the unrolled code is very slow to compile, especially if all architectures are enabled as they are by default.