@JohannesGaessler 2023-05-21
I ended up doing an implementation kind of similar to what ggerganov proposed. It's now possible to set an option `LLAMA_CUDA_BX` that sets the how many data values in x direction are processed in each outer loop iteration. This then also increases the amount of data processed in the unrolled inner loop. The optimal parameters for my RTX 3090 are `LLAMA_CUDA_BX=64 LLAMA_CUDA_BY=2` which result in 14.53 t/s for 33b q4_0.

In terms of features I think this PR is complete but I just noticed that "BLOCK_X" is maybe a bad name for this value since it does not actually set the block size; that value is always `WARP SIZE` = 32. Suggestions for better naming?