@SlyEcho 2023-05-25
This is amazing now, <70 ms/t for 13b Q4_0 on my old graphics card.

The variables need to be tuned though for different systems.